refactor(gemini): extract shared getBunGeminiPath utility
Extract duplicate getBunGeminiPath() function to a shared utility
module to improve code maintainability and fix code review issues.

Changes:
- Create cli/src/gemini/utils/bunGeminiPath.ts with shared function
- Update geminiLocal.ts to use shared utility
- Update geminiBackend.ts to use shared utility
- Change logger.info to logger.debug for less verbose output
- Add comprehensive JSDoc documentation
- Unify error handling with consistent logging

Benefits:
- Eliminates 63 lines of duplicate code
- Easier to maintain and update Bun detection logic
- Consistent error logging across both use cases
- Better code documentation for future contributors

Addresses code review findings for PR #128
