+++ cli/src/gemini/geminiLocal.ts
@@ -1,38 +1,7 @@
 import { logger } from '@/ui/logger';
 import { restoreTerminalState } from '@/ui/terminalState';
 import { spawnWithAbort } from '@/utils/spawnWithAbort';
-import { existsSync } from 'node:fs';
-import { homedir } from 'node:os';
-import { join } from 'node:path';
-import { spawnSync } from 'node:child_process';
-
-/**
- * Check if Bun-optimized Gemini CLI is available
- * Returns the path if available, null otherwise
- */
-function getBunGeminiPath(): string | null {
-    try {
-        const bunGeminiPath = join(homedir(), '.bun', 'install', 'global', 'node_modules', '@google', 'gemini-cli', 'dist', 'index.js');
-
-        // Check if Bun version exists
-        if (!existsSync(bunGeminiPath)) {
-            return null;
-        }
-
-        // Check if bun command is available
-        const bunCheck = spawnSync('bun', ['--version'], { stdio: 'ignore' });
-
-        if (bunCheck.error || bunCheck.status !== 0) {
-            logger.debug('[GeminiLocal] Bun command not available, falling back to standard Gemini CLI');
-            return null;
-        }
-
-        return bunGeminiPath;
-    } catch (error) {
-        logger.debug('[GeminiLocal] Error checking Bun availability:', error);
-        return null;
-    }
-}
+import { getBunGeminiPath } from './utils/bunGeminiPath';
 
 export async function geminiLocal(opts: {
     path: string;
@@ -72,8 +41,8 @@ export async function geminiLocal(opts: {
     process.stdin.pause();
     try {
         if (bunGeminiPath) {
-            // Use Bun runtime for faster startup
-            logger.info('[GeminiLocal] Using Bun-optimized Gemini CLI (faster startup)');
+            // Use Bun runtime for faster startup (~2x faster than Node.js)
+            logger.debug('[GeminiLocal] Using Bun-optimized Gemini CLI (faster startup)');
             await spawnWithAbort({
                 command: 'bun',
                 args: ['run', bunGeminiPath, ...args],
@@ -89,7 +58,7 @@ export async function geminiLocal(opts: {
             });
         } else {
             // Fallback to original gemini command (non-breaking)
-            logger.info('[GeminiLocal] Using standard Gemini CLI (fallback mode)');
+            logger.debug('[GeminiLocal] Using standard Gemini CLI (fallback mode)');
             await spawnWithAbort({
                 command: 'gemini',
                 args,
