+++ cli/src/gemini/utils/geminiBackend.ts
@@ -1,9 +1,6 @@
 import { AcpSdkBackend } from '@/agent/backends/acp';
 import { buildGeminiEnv, resolveGeminiRuntimeConfig } from './config';
-import { existsSync } from 'node:fs';
-import { homedir } from 'node:os';
-import { join } from 'node:path';
-import { spawnSync } from 'node:child_process';
+import { getBunGeminiPath } from './bunGeminiPath';
 
 function filterEnv(env: NodeJS.ProcessEnv): Record<string, string> {
     const result: Record<string, string> = {};
@@ -13,33 +10,6 @@ function filterEnv(env: NodeJS.ProcessEnv): Record<string, string> {
         }
     }
     return result;
-}
-
-/**
- * Check if Bun-optimized Gemini CLI is available
- * Returns the path if both Bun and Bun-optimized Gemini are available, null otherwise
- */
-function getBunGeminiPath(): string | null {
-    try {
-        const bunGeminiPath = join(homedir(), '.bun', 'install', 'global', 'node_modules', '@google', 'gemini-cli', 'dist', 'index.js');
-
-        // Check if Bun version of Gemini CLI exists
-        if (!existsSync(bunGeminiPath)) {
-            return null;
-        }
-
-        // Check if bun command itself is available on the system
-        const bunCheck = spawnSync('bun', ['--version'], { stdio: 'ignore' });
-
-        if (bunCheck.error || bunCheck.status !== 0) {
-            return null;
-        }
-
-        return bunGeminiPath;
-    } catch {
-        // If any error occurs during check, safely return null
-        return null;
-    }
 }
 
 export function createGeminiBackend(opts: {
